requirements.txt
attrs==22.1.0
azure-core==1.26.1
azure-identity==1.12.0
azure-storage-blob==12.14.1
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
confection==0.0.3
cryptography==38.0.4
cymem==2.0.7
docx2txt==0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl
idna==3.4
isodate==0.6.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
langcodes==3.3.0
MarkupSafe==2.1.1
msal==1.20.0
msal-extensions==1.0.0
msrest==0.7.1
murmurhash==1.0.9
nltk==3.8
numpy==1.23.5
oauthlib==3.2.2
packaging==22.0
pandas==1.5.2
pathy==0.10.1
pdfminer.six==20221105
portalocker==2.6.0
preshed==3.0.8
pycparser==2.21
pycryptodome==3.16.0
pydantic==1.10.2
PyJWT==2.6.0
pyresparser==1.0.6
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
smart-open==6.3.0
sortedcontainers==2.4.0
spacy==3.4.4
spacy-legacy==3.0.10
spacy-loggers==1.0.4
srsly==2.4.5
thinc==8.1.5
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.13
wasabi==0.10.1
